+++ Split.pde
@@ -2,10 +2,10 @@ public class Split {
   int x;
   int y;
   boolean isActive;
-  float pressure;
+  int pressure;
   
   
-  public Split (int x, int y, float pressure) {
+  public Split (int x, int y, int pressure) {
     this.x = x;
     this.y = y;
     this.isActive = true ;
